requirements.txt
amqp==2.6.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
asgiref==3.5.0; python_version >= '3.7'
async-timeout==4.0.2; python_version >= '3.6'
bcrypt==3.2.0
billiard==3.6.4.0
brotli==1.0.9
cachetools==5.0.0; python_version ~= '3.7'
carbonize==0.0.1
celery==4.4.7
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12; python_version >= '3'
cloudscraper==1.2.60
commonmark==0.9.1
coreapi==2.3.3
coreschema==0.0.4
cryptography==36.0.2
cssselect2==0.5.0; python_version >= '3.7'
cssselect==1.1.0
defusedxml==0.7.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
deprecated==1.2.13; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
django-allauth==0.50.0
django-anymail==8.5
django-appconf==1.0.5; python_version >= '3.6'
django-compression-middleware==0.4.1
django-compressor==3.1
django-cors-headers==3.11.0
django-countries==7.3.2
django-crispy-forms==1.14.0
django-dnt==1.0
django-libsass==0.9
django-recaptcha==3.0.0
django==4.0.3
djangorestframework-csv==2.1.1
djangorestframework==3.13.1
drf-writable-nested==0.6.3
environs==9.5.0
et-xmlfile==1.1.0; python_version >= '3.6'
gevent==21.12.0
google-api-core[grpc]==2.7.1; python_version >= '3.6'
google-auth==2.6.2; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
google-cloud-language==2.4.1
googleapis-common-protos==1.56.0; python_version >= '3.6'
greenlet==1.1.2; platform_python_implementation == 'CPython'
grpcio-status==1.45.0
grpcio==1.45.0
gunicorn==20.1.0
hiredis==2.0.0
huey==2.4.3
idna==3.3; python_version >= '3'
itypes==1.2.0
jinja2==3.1.1; python_version >= '3.7'
kombu==4.6.11; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
libsass==0.21.0
lxml==4.8.0
markupsafe==2.1.1; python_version >= '3.7'
marshmallow==3.15.0; python_version >= '3.7'
oauthlib==3.2.0; python_version >= '3.6'
openpyxl==3.0.9
packaging==21.3; python_version >= '3.6'
pillow==9.1.0
plotly==5.6.0
proto-plus==1.20.3; python_version >= '3.6'
protobuf==3.20.0; python_version >= '3.7'
psycopg2==2.9.3
pyasn1-modules==0.2.8
pyasn1==0.4.8
pycparser==2.21
pyjwt[crypto]==2.3.0; python_version >= '3.6'
pyparsing==3.0.7; python_version >= '3.6'
python-dotenv==0.20.0; python_version >= '3.5'
python3-openid==3.2.0
pytz==2022.1
rcssmin==1.1.0
redis==4.2.2
reportlab==3.6.9
requests-oauthlib==1.3.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
requests-toolbelt==0.9.1
requests==2.27.1
rjsmin==1.2.0
rsa==4.8; python_version >= '3.6'
sentry-sdk==1.5.8
setuptools==62.0.0; python_version >= '3.7'
six==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
sqlparse==0.4.2; python_version >= '3.5'
svglib==1.2.1
tenacity==8.0.1; python_version >= '3.6'
tinycss2==1.1.1; python_version >= '3.6'
twitter==1.19.3
typing-extensions==4.1.1; python_version >= '3.6'
unicodecsv==0.14.1
uritemplate==4.1.1; python_version >= '3.6'
urllib3==1.26.9; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4' and python_version < '4'
vine==1.3.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
webencodings==0.5.1
whitenoise==6.0.0
wrapt==1.14.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
zoomus==1.1.6
zope.event==4.5.0
zope.interface==5.4.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
zstandard==0.17.0; python_version >= '3.6'
